refactor(StringSplitter): util에서 service 패키지로 이동
- StringSplitter 클래스를 util 패키지에서 service 패키지로 이동
